+++ src/plugins/dam/components/asset-snippet/Image.jsx
@@ -23,46 +23,48 @@ const Image = ({ asset, previewUrl }) => (
       </FormGroup>
     </div>
     <div className="col-12 col-md-6">
-      <div className="dam-media-container">
-        <ReactImageMagnify
-          smallImage={{
-            src: previewUrl || damUrl(asset, 'o', 'sm'),
-            isFluidWidth: true,
-          }}
-          imageStyle={{
-            objectFit: 'contain',
-          }}
-          largeImage={{
-            src: previewUrl || damUrl(asset, 'o'),
-            width: (asset.get('width')),
-            height: (asset.get('height')),
-          }}
-          enlargedImageContainerStyle={{
-            position: 'fixed',
-            top: '50%',
-            left: '50%',
-            border: '0px',
-            marginLeft: 0,
-            maxHeight: '100vh',
-            maxWidth: '100vw',
-            transform: 'translate(-50%, -50%)',
-            zIndex: 1049,
-            pointerEvents: 'none',
-          }}
-          enlargedImagePosition="beside"
-          enlargedImageContainerDimensions={{ width: '100%' }}
-          enlargedImageContainerClassName="w-auto"
-          enlargedImageStyle={{
-            margin: '16px',
-            border: '0px',
-            maxHeight: 'calc(90vh)',
-            maxWidth: 'calc(90vw)',
-            filter: 'drop-shadow(0 1px 4px rgba(0,0,0,0.4)',
-            objectFit: 'scale-down',
-          }}
-          enlargedImageClassName="h-100 w-auto asset-hover-mw"
-        />
-      </div>
+      {previewUrl && (
+        <div className="dam-media-container">
+          <ReactImageMagnify
+            smallImage={{
+              src: previewUrl || damUrl(asset, 'o', 'sm'),
+              isFluidWidth: true,
+            }}
+            imageStyle={{
+              objectFit: 'contain',
+            }}
+            largeImage={{
+              src: previewUrl || damUrl(asset, 'o'),
+              width: (asset.get('width')),
+              height: (asset.get('height')),
+            }}
+            enlargedImageContainerStyle={{
+              position: 'fixed',
+              top: '50%',
+              left: '50%',
+              border: '0px',
+              marginLeft: 0,
+              maxHeight: '100vh',
+              maxWidth: '100vw',
+              transform: 'translate(-50%, -50%)',
+              zIndex: 1049,
+              pointerEvents: 'none',
+            }}
+            enlargedImagePosition="beside"
+            enlargedImageContainerDimensions={{ width: '100%' }}
+            enlargedImageContainerClassName="w-auto"
+            enlargedImageStyle={{
+              margin: '16px',
+              border: '0px',
+              maxHeight: 'calc(90vh)',
+              maxWidth: 'calc(90vw)',
+              filter: 'drop-shadow(0 1px 4px rgba(0,0,0,0.4)',
+              objectFit: 'scale-down',
+            }}
+            enlargedImageClassName="h-100 w-auto asset-hover-mw"
+          />
+        </div>
+      )}
     </div>
   </div>
 );