+++ src/plugins/notify/components/search-notifications-form/index.jsx
@@ -91,17 +91,14 @@ class SearchNotificationsForm extends Component {
   handleChangeStatus(selectedOption) {
     const { request } = this.state;
     const searchStatus = get(selectedOption, 'value', null);
-
+    const status = (searchStatus === NotificationSendStatus.CANCELED.toString()) ? NodeStatus.DELETED : null;
     const requestData = {
       request: {
         ...request,
         send_status: searchStatus,
+        status,
       },
     };
-
-    if (searchStatus === NotificationSendStatus.CANCELED.toString()) {
-      requestData.request.status = NodeStatus.DELETED;
-    }
 
     this.setState(requestData, this.handleCollapseOpen);
   }