+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 
 
+## patch
+* fix node type filtering is not sticky when leaving and re-entering the page
+
+
 ## v0.2.9
 * Correct `isClearable` behavior for `Select` components.