+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## patch
 * fix node type filtering is not sticky when leaving and re-entering the page
+* Do not replace trailing space with a dash when constructing slugs
 
 
 ## v0.2.9