+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # CHANGELOG
 
 
-## vN.N.N
+## patch
+* Fix node type filtering is not sticky when leaving and re-entering the page
+* Do not replace trailing space with a dash when constructing slugs
+* Correct `isClearable` behavior for `Select` components.
+* Correct `pbjUrl` template for poll and widget pickers.
+* Fix structure > teasers if pressing 1 to change order date, the selector/cursor will lose focus.
+* Add `autocomplete: true` to the search nodes request in node picker field.
 * Add `UpdateNodeEnricher.js` which will add all paths (aka fields) that have changed during the `gdbots:ncr:mixin:update-node.enrich` lifecycle event.